Split out compile_commands.json handling into library and add a generic compdb extractor
