+++ kythe/go/extractors/config/runextractor/cmakecmd/cmakecmd.go
@@ -19,23 +19,18 @@ package cmakecmd
 
 import (
 	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strings"
 
+	"kythe.io/kythe/go/extractors/config/runextractor/compdb"
 	"kythe.io/kythe/go/util/cmdutil"
 
-	"bitbucket.org/creachadair/shell"
 	"github.com/google/subcommands"
 	"github.com/pborman/uuid"
-	"golang.org/x/sync/errgroup"
-	"golang.org/x/sync/semaphore"
 )
 
 type cmakeCommand struct {
@@ -44,12 +39,6 @@ type cmakeCommand struct {
 	extractor string
 	buildDir  string
 	sourceDir string
-}
-
-// A compileCommand holds the decoded arguments of a CMake JSON command spec.
-type compileCommand struct {
-	Command   string
-	Directory string
 }
 
 // New creates a new subcommand for running cmake extraction.
@@ -126,8 +115,7 @@ func (c *cmakeCommand) Execute(ctx context.Context, fs *flag.FlagSet, args ...in
 		return c.Fail("error building repository: %v", err)
 	}
 
-	// TODO(shahms): Move compile_commands.json handling to common library.
-	if err := extractCompilations(ctx, extractor, filepath.Join(buildDir, "compile_commands.json")); err != nil {
+	if err := compdb.ExtractCompilations(ctx, extractor, filepath.Join(buildDir, "compile_commands.json")); err != nil {
 		return c.Fail("error extracting repository: %v", err)
 	}
 	return subcommands.ExitSuccess
@@ -140,89 +128,8 @@ func cleanBuild(dir string) {
 	}
 }
 
-// readCommands reads the JSON file at path into a slice of compileCommands.
-func readCommands(path string) ([]compileCommand, error) {
-	data, err := ioutil.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-	var commands []compileCommand
-	if err := json.Unmarshal(data, &commands); err != nil {
-		return nil, err
-	}
-	return commands, nil
-
-}
-
-// extractorEnv copies the existing environment and modifies it to be suitable for an extractor invocation.
-func extractorEnv() ([]string, error) {
-	var env []string
-	for _, value := range os.Environ() {
-		parts := strings.SplitN(value, "=", 2)
-		// Until kzip support comes along, we only support writing to a single directory so strip these options.
-		if parts[0] == "KYTHE_INDEX_PACK" || parts[0] == "KYTHE_OUTPUT_FILE" {
-			continue
-		} else if parts[0] == "KYTHE_OUTPUT_DIRECTORY" {
-			// Remap KYTHE_OUTPUT_DIRECTORY to be an absolute path.
-			output, err := filepath.Abs(parts[1])
-			if err != nil {
-				return nil, err
-			}
-			env = append(env, "KYTHE_OUTPUT_DIRECTORY="+output)
-		} else {
-			// Otherwise, preserve the environment unchanged.
-			env = append(env, value)
-		}
-
-	}
-	return env, nil
-}
-
 // runIn changes the cmd working directory to dir and call Run.
 func runIn(cmd *exec.Cmd, dir string) error {
 	cmd.Dir = dir
 	return cmd.Run()
-}
-
-// extractCompilations runs the extractor over each record from the compile_commands.json file at path.
-func extractCompilations(ctx context.Context, extractor, path string) error {
-	commands, err := readCommands(path)
-	if err != nil {
-		return err
-	}
-	exEnv, err := extractorEnv()
-	if err != nil {
-		return err
-	}
-	sem := semaphore.NewWeighted(128) // Limit concurrency.
-	extraction, ctx := errgroup.WithContext(ctx)
-	for _, entry := range commands {
-		entry := entry
-		extraction.Go(func() error {
-			if err := sem.Acquire(ctx, 1); err != nil {
-				return err
-			}
-			defer sem.Release(1)
-
-			cmd := exec.CommandContext(ctx, extractor, "--with_executable")
-			args, ok := shell.Split(entry.Command)
-			if !ok {
-				return fmt.Errorf("unable to split command line")
-			}
-			cmd.Args = append(cmd.Args, args...)
-			cmd.Dir, err = filepath.Abs(entry.Directory)
-			if err != nil {
-				return fmt.Errorf("unable to resolve cmake directory: %v", err)
-			}
-			cmd.Env = exEnv
-			if _, err := cmd.Output(); err != nil {
-				if exit, ok := err.(*exec.ExitError); ok {
-					return fmt.Errorf("error running extractor: %v (%s)", exit, exit.Stderr)
-				}
-				return fmt.Errorf("error running extractor: %v", err)
-			}
-			return nil
-		})
-	}
-	return extraction.Wait()
 }
